layout/style/test/test_units_angle.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for serialization and equivalence of angle units</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=">Mozilla Bug </a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for serialization and equivalence of angle units **/

/**
 * We test that for each of the following:
 *  + they reserialize to exactly what is given
 *  + if a mapping is provided, they compute to the same result as the mapping
 */
var tests = {
  "45deg": "50grad",
  "150grad": "135deg",
  "1rad": null
};

var p = document.getElementById("display");

for (var test in tests) {
  p.setAttribute("style", "-moz-transform: rotate(" + test + ")");
  is(p.style.getPropertyValue("-moz-transform"), "rotate(" + test + ")",
     test + " serializes to exactly itself");
  // We can't test any equivalence since we don't have any properties
  // with angle values that we compute.  (-moz-transform doesn't help.)
/*
  var equiv = tests[test];
  if (equiv) {
    var cm1 = getComputedStyle(p, "").elevation;
    p.style.elevation = equiv;
    var cm2 = getComputedStyle(p, "").elevation;
    is(cm1, cm2, test + " should compute to the same as " + equiv);
  }
*/
}

</script>
</pre>
</body>
</html>
